Theme_3_2_API/Theme_3_2_new_PRO/requirements.txt
psycopg2-binary==2.9.6
pandas==1.5.3
requests==2.28.1